Improve wallet interface
- Separate error variants for wallet and Vault
- Move wallet code into its own module
